+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Upgrade the `org.realityforge.arez` artifacts to version `0.173`.
+
 ### [v6.83](https://github.com/replicant4j/replicant/tree/v6.83) (2020-04-02) · [Full Changelog](https://github.com/replicant4j/replicant/compare/v6.82...v6.83)
 
 * Change the routing infrastructure so an entity message can be associated with multiple instance graphs of the same type. i.e. An entity can belong to the `Event/1` and `Event/2` channel.